server/requirements.txt
attrs==22.1.0
dnspython==2.2.1
docopt==0.6.2
exceptiongroup==1.0.1
iniconfig==1.1.1
mpmath==1.2.1
numpy==1.23.4
packaging==21.3
pluggy==1.0.0
pycryptodome==3.15.0
pydantic==1.10.2
pymongo==4.3.2
pyparsing==3.0.9
pytest==7.2.0
sympy==1.11.1
tinyec==0.4.0
tomli==2.0.1
typing_extensions==4.4.0
